pyproject.toml
[tool.poetry]
name = "music-recog"
version = "0.1.0"
description = ""
authors = ["James Baskerville <jameswbaskerville@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
aiortc = {extras = ["standard"], version = "^1.9.0"}
python-dotenv = "^0.19.0"
aiohttp = "^3.8.1"
sse-starlette = "^2.1.3"
python-multipart = "^0.0.12"
httpx = "^0.27.2"
numpy = "^2.1.2"
pydub = {extras = ["standard"], version = "^0.25.1"}
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
